Create LICENSE.md
I want a prettier version than LICENSE a .md file would be nice, but let's see how this looks in practice okay? Maybe there's a way to link the CC's graphical version of the license here? And why isn't the CC-BY-SA or any other CC license other than the CC-0 available from the drop down list like it is in Gitea?